Removed Secure Var
I know what you're all going to say... but look it's a dedicated test key just reading data so I don't see what harm it could do by it being out there. Someone could steal it and use it to do something malicious sure... but the alternative is that we can't build PR's from other repos which is just shit. I'll cycle it every week or so to be sure.